backend/requirements.txt
asgiref==3.3.1
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.4
defusedxml==0.5.0
Django==2.2.13
django-allauth==0.37.1
django-appconf==1.0.4
django-cors-headers==2.4.0
django-imagekit==4.0.2
django-rest-auth==0.9.3
djangorestframework==3.9.1
gunicorn==19.7.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.20.1
oauthlib==2.1.0
opencv-python==4.5.1.48
pilkit==2.0
Pillow==8.1.0
pylint==2.1.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.3.0
urllib3==1.24.2
wrapt==1.10.11
